+++ scripts/build-react.js
@@ -1,14 +1,13 @@
+import execSh from 'exec-sh';
+import { outputDir } from './utils/output-dir.js';
+import { addBannerToFile } from './utils/banner.js';
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 /* eslint no-console: "off" */
-const exec = require('exec-sh').promise;
-const { outputDir } = require('./utils/output-dir');
-const { addBannerToFile } = require('./utils/banner');
+const exec = execSh.promise;
 
-async function buildReact() {
+export default async function buildReact() {
   await exec(
-    `npx babel --config-file ./scripts/babel/babel.config.react.js src/react --out-dir ${outputDir}/react`,
+    `npx babel --config-file ./scripts/babel/babel.config.react.json src/react --out-dir ${outputDir}/react`,
   );
   await addBannerToFile(`./${outputDir}/react/swiper-react.js`, 'React');
 }
-
-module.exports = buildReact;